parseFilterOptions moved to helperFunctions from postService and countryService
