+++ server/src/services/postService.ts
@@ -5,6 +5,8 @@ import BadRequestError from '../errors/BadRequestError';
 import { UserModel } from '../model/UserModel';
 import { FollowModel } from '../model/FollowModel';
 import { LikeModel } from '../model/LikeModel';
+import { parseFilterOptions } from './helperFunctions';
+
 
 export const getPostsFromFollowing = async (
     userId: Types.ObjectId,
@@ -195,47 +197,7 @@ export const updatePost = async (
     return updatedPost;
 };
 
-const parseFilterOptions = (options: {
-    page?: string;
-    limit?: string;
-    search?: string;
-    sort?: string;
-    asc?: string;
-}) => {
-    const {
-        page = '1',
-        limit = '10',
-        search = '',
-        sort = 'created_at',
-        asc = 'false',
-    } = options;
 
-    const parsedPage = parseInt(page, 10);
-    const parsedLimit = parseInt(limit, 10);
-    const parsedAsc = asc === 'false';
-
-    if (isNaN(parsedPage) || parsedPage < 1) {
-        throw new BadRequestError({
-            code: 400,
-            message: 'Invalid page number. Must be a positive integer.',
-        });
-    }
-
-    if (isNaN(parsedLimit) || parsedLimit < 1) {
-        throw new BadRequestError({
-            code: 400,
-            message: 'Invalid limit. Must be a positive integer.',
-        });
-    }
-
-    return {
-        page: parsedPage,
-        limit: parsedLimit,
-        search,
-        sort,
-        asc: parsedAsc,
-    };
-};
 
 export const filterPosts = async (options: {
     page?: string;
